+++ libft/printf/fpf_conversions.c
@@ -12,44 +12,44 @@
 
 #include "ft_printf.h"
 
-signed long long int	fpf_conv_sig_int(t_printf *fpf)
+intmax_t	fpf_conv_sig_int(t_printf *fpf)
 {
-	signed long long int num;
+	intmax_t num;
 
 	if (fpf->size_flag == SF_HH)
-		num = (signed char)va_arg(fpf->args, signed long long int);
+		num = (signed char)va_arg(fpf->args, intmax_t);
 	else if (fpf->size_flag == SF_H)
-		num = (signed short)va_arg(fpf->args, signed long long int);
+		num = (signed short)va_arg(fpf->args, intmax_t);
 	else if (fpf->size_flag == SF_L)
-		num = (signed long)va_arg(fpf->args, signed long long int);
+		num = (signed long)va_arg(fpf->args, intmax_t);
 	else if (fpf->size_flag == SF_LL)
-		num = (signed long long)va_arg(fpf->args, signed long long int);
+		num = (signed long long)va_arg(fpf->args, intmax_t);
 	else if (fpf->size_flag == SF_J)
-		num = (intmax_t)va_arg(fpf->args, signed long long int);
+		num = (intmax_t)va_arg(fpf->args, intmax_t);
 	else if (fpf->size_flag == SF_Z)
-		num = (size_t)va_arg(fpf->args, signed long long int);
+		num = (size_t)va_arg(fpf->args, intmax_t);
 	else
-		num = (signed int)va_arg(fpf->args, signed long long int);
+		num = (signed int)va_arg(fpf->args, intmax_t);
 	return (num);
 }
 
-unsigned long long int	fpf_conv_unsig_int(t_printf *fpf)
+uintmax_t	fpf_conv_unsig_int(t_printf *fpf)
 {
-	unsigned long long int num;
+	uintmax_t num;
 
 	if (fpf->size_flag == SF_HH)
-		num = (unsigned char)va_arg(fpf->args, unsigned long long int);
+		num = (unsigned char)va_arg(fpf->args, uintmax_t);
 	else if (fpf->size_flag == SF_H)
-		num = (unsigned short)va_arg(fpf->args, unsigned long long int);
+		num = (unsigned short)va_arg(fpf->args, uintmax_t);
 	else if (fpf->size_flag == SF_L)
-		num = (unsigned long)va_arg(fpf->args, unsigned long long int);
+		num = (unsigned long)va_arg(fpf->args, uintmax_t);
 	else if (fpf->size_flag == SF_LL)
-		num = (unsigned long long)va_arg(fpf->args, unsigned long long int);
+		num = (unsigned long long)va_arg(fpf->args, uintmax_t);
 	else if (fpf->size_flag == SF_J)
-		num = (uintmax_t)va_arg(fpf->args, unsigned long long int);
+		num = (uintmax_t)va_arg(fpf->args, uintmax_t);
 	else if (fpf->size_flag == SF_Z)
-		num = (size_t)va_arg(fpf->args, unsigned long long int);
+		num = (size_t)va_arg(fpf->args, uintmax_t);
 	else
-		num = (unsigned int)va_arg(fpf->args, unsigned long long int);
+		num = (unsigned int)va_arg(fpf->args, uintmax_t);
 	return (num);
 }
